[MSLabel] Manually merging in Alexefish's pull request. Made some modifications to keep consistent style. Copying his notes here:
The new line calculation was taking into account an extra space where it did not need to. Therefore a linebreak would be inserted too early with some strings

if ([[line stringByAppendingFormat:@"%@ ", word] sizeWithFont:self.font].width <= self.frame.size.width)

Was calculating the string width with an extra space, updated it to:

if ([[line stringByAppendingFormat:@"%@", word] sizeWithFont:self.font].width <= self.frame.size.width)

So that new line breaks are not inserted too early for the string, whole words are still maintained on a new line.